Split code actions into several files
