+++ tooling/lsp/src/requests/code_action.rs
@@ -7,27 +7,24 @@ use async_lsp::ResponseError;
 use fm::{FileId, FileMap, PathString};
 use lsp_types::{
     CodeAction, CodeActionKind, CodeActionOrCommand, CodeActionParams, CodeActionResponse,
-    Position, Range, TextDocumentPositionParams, TextEdit, Url, WorkspaceEdit,
+    TextDocumentPositionParams, TextEdit, Url, WorkspaceEdit,
 };
-use noirc_errors::{Location, Span};
+use noirc_errors::Span;
 use noirc_frontend::{
-    ast::{ConstructorExpression, Ident, Path, Visitor},
+    ast::{ConstructorExpression, Path, Visitor},
     graph::CrateId,
     hir::def_map::{CrateDefMap, LocalModuleId, ModuleId},
-    macros_api::{ModuleDefId, NodeInterner},
-    node_interner::ReferenceId,
+    macros_api::NodeInterner,
     parser::{Item, ItemKind, ParsedSubModule},
     ParsedModule,
 };
 
-use crate::{
-    byte_span_to_range,
-    modules::{get_parent_module_id, module_full_path},
-    utils, LspState,
-};
+use crate::{utils, LspState};
 
 use super::{process_request, to_lsp_location};
 
+mod fill_struct_fields;
+mod import_or_qualify;
 #[cfg(test)]
 mod tests;
 
@@ -140,45 +137,6 @@ impl<'a> CodeActionFinder<'a> {
         Some(code_actions)
     }
 
-    fn push_import_code_action(&mut self, full_path: &str) {
-        let line = self.auto_import_line as u32;
-        let character = (self.nesting * 4) as u32;
-        let indent = " ".repeat(self.nesting * 4);
-        let mut newlines = "\n";
-
-        // If the line we are inserting into is not an empty line, insert an extra line to make some room
-        if let Some(line_text) = self.lines.get(line as usize) {
-            if !line_text.trim().is_empty() {
-                newlines = "\n\n";
-            }
-        }
-
-        let title = format!("Import {}", full_path);
-        let text_edit = TextEdit {
-            range: Range { start: Position { line, character }, end: Position { line, character } },
-            new_text: format!("use {};{}{}", full_path, newlines, indent),
-        };
-
-        let code_action = self.new_quick_fix(title, text_edit);
-        self.code_actions.push(CodeActionOrCommand::CodeAction(code_action));
-    }
-
-    fn push_qualify_code_action(&mut self, ident: &Ident, prefix: &str, full_path: &str) {
-        let Some(range) = byte_span_to_range(
-            self.files,
-            self.file,
-            ident.span().start() as usize..ident.span().start() as usize,
-        ) else {
-            return;
-        };
-
-        let title = format!("Qualify as {}", full_path);
-        let text_edit = TextEdit { range, new_text: format!("{}::", prefix) };
-
-        let code_action = self.new_quick_fix(title, text_edit);
-        self.code_actions.push(CodeActionOrCommand::CodeAction(code_action));
-    }
-
     fn new_quick_fix(&self, title: String, text_edit: TextEdit) -> CodeAction {
         let mut changes = HashMap::new();
         changes.insert(self.uri.clone(), vec![text_edit]);
@@ -247,58 +205,7 @@ impl<'a> Visitor for CodeActionFinder<'a> {
     }
 
     fn visit_path(&mut self, path: &Path) {
-        if path.segments.len() != 1 {
-            return;
-        }
-
-        let ident = &path.segments[0].ident;
-        if !self.includes_span(ident.span()) {
-            return;
-        }
-
-        let location = Location::new(ident.span(), self.file);
-        if self.interner.find_referenced(location).is_some() {
-            return;
-        }
-
-        let current_module_parent_id = get_parent_module_id(self.def_maps, self.module_id);
-
-        // The Path doesn't resolve to anything so it means it's an error and maybe we
-        // can suggest an import or to fully-qualify the path.
-        for (name, entries) in self.interner.get_auto_import_names() {
-            if name != &ident.0.contents {
-                continue;
-            }
-
-            for (module_def_id, visibility) in entries {
-                let Some(module_full_path) = module_full_path(
-                    *module_def_id,
-                    *visibility,
-                    self.module_id,
-                    current_module_parent_id,
-                    self.interner,
-                ) else {
-                    continue;
-                };
-
-                let full_path = if let ModuleDefId::ModuleId(..) = module_def_id {
-                    module_full_path.clone()
-                } else {
-                    format!("{}::{}", module_full_path, name)
-                };
-
-                let qualify_prefix = if let ModuleDefId::ModuleId(..) = module_def_id {
-                    let mut segments: Vec<_> = module_full_path.split("::").collect();
-                    segments.pop();
-                    segments.join("::")
-                } else {
-                    module_full_path
-                };
-
-                self.push_import_code_action(&full_path);
-                self.push_qualify_code_action(ident, &qualify_prefix, &full_path);
-            }
-        }
+        self.import_or_qualify(path);
     }
 
     fn visit_constructor_expression(
@@ -310,103 +217,7 @@ impl<'a> Visitor for CodeActionFinder<'a> {
             return false;
         }
 
-        // Find out which struct this is
-        let location = Location::new(constructor.type_name.last_ident().span(), self.file);
-        let Some(ReferenceId::Struct(struct_id)) = self.interner.find_referenced(location) else {
-            return true;
-        };
-
-        let struct_type = self.interner.get_struct(struct_id);
-        let struct_type = struct_type.borrow();
-
-        // First get all of the struct's fields
-        let mut fields = struct_type.get_fields_as_written();
-
-        // Remove the ones that already exists in the constructor
-        for (field, _) in &constructor.fields {
-            fields.retain(|(name, _)| name != &field.0.contents);
-        }
-
-        if fields.is_empty() {
-            return true;
-        }
-
-        // Some fields are missing. Let's suggest a quick fix that adds them.
-        let bytes = self.source.as_bytes();
-        let right_brace_index = span.end() as usize - 1;
-        let mut index = right_brace_index - 1;
-        while bytes[index].is_ascii_whitespace() {
-            index -= 1;
-        }
-
-        let char_before_right_brace = bytes[index] as char;
-
-        index += 1;
-
-        let Some(range) = byte_span_to_range(self.files, self.file, index..index) else {
-            return true;
-        };
-
-        // If the constructor spans multiple lines, we'll add the new fields in new lines too.
-        // Otherwise we'll add all the fields in a single line.
-        let constructor_range =
-            byte_span_to_range(self.files, self.file, span.start() as usize..span.end() as usize);
-
-        // If it's multiline, find out the indent of the beginning line: we'll add new fields
-        // with that indent "plus one" (4 more spaces).
-        let line_indent = if let Some(constructor_range) = constructor_range {
-            if constructor_range.start.line == constructor_range.end.line {
-                None
-            } else {
-                let line = self.lines[constructor_range.start.line as usize];
-                let whitespace_bytes =
-                    line.bytes().take_while(|byte| byte.is_ascii_whitespace()).count();
-                Some(whitespace_bytes)
-            }
-        } else {
-            None
-        };
-        let line_indent = line_indent.map(|indent| " ".repeat(indent + 4));
-
-        let on_whitespace = bytes[index].is_ascii_whitespace();
-
-        let mut new_text = String::new();
-
-        // Add a comma if there's not a trailing one (if there are existing fields)
-        if !constructor.fields.is_empty() && char_before_right_brace != ',' {
-            new_text.push(',');
-        }
-
-        // Add space or newline depending on whether it's multiline or not
-        if let Some(line_indent) = &line_indent {
-            new_text.push('\n');
-            new_text.push_str(line_indent);
-        } else if !on_whitespace || constructor.fields.is_empty() {
-            new_text.push(' ');
-        }
-
-        for (index, (name, _)) in fields.iter().enumerate() {
-            if index > 0 {
-                new_text.push(',');
-                if let Some(line_indent) = &line_indent {
-                    new_text.push('\n');
-                    new_text.push_str(line_indent);
-                } else {
-                    new_text.push(' ');
-                }
-            }
-            new_text.push_str(name);
-            new_text.push_str(": ()");
-        }
-
-        if !bytes[right_brace_index - 1].is_ascii_whitespace() {
-            new_text.push(' ');
-        }
-
-        let title = "Fill struct fields".to_string();
-        let text_edit = TextEdit { range, new_text };
-        let code_action = self.new_quick_fix(title, text_edit);
-        self.code_actions.push(CodeActionOrCommand::CodeAction(code_action));
+        self.fill_struct_fields(constructor, span);
 
         true
     }
